Testes separados por arquivo .h Agora cada modulo do sistema possui um arquivo de teste que pode ser executado sozinho ou simultaneamente com outro.
